requirements.txt
openai==0.27.2
Flask==2.3.2
numpy==1.24.3
pandas==2.0.1
requests==2.30.0
joblib==1.2.0
python-dotenv==1.0.0
scikit-learn==1.2.2

